fix(cleanup)
Cleaned code
Improved doc
